clean up webpack config and make things more dry
